# Use unsafe operations to get fast memory copy