--- a/pkg/api/v1/defaults.go
+++ b/pkg/api/v1/defaults.go
@@ -52,7 +52,19 @@ func addDefaultingFuncs(scheme *runtime.Scheme) error {
 		SetDefaults_LimitRangeItem,
 		SetDefaults_ConfigMap,
 		SetDefaults_RBDVolumeSource,
+		SetDefaults_ResourceList,
 	)
+}
+
+func SetDefaults_ResourceList(obj *ResourceList) {
+	for key, val := range *obj {
+		// TODO(#18538): We round up resource values to milli scale to maintain API compatibility.
+		// In the future, we should instead reject values that need rounding.
+		const milliScale = -3
+		val.RoundUp(milliScale)
+
+		(*obj)[ResourceName(key)] = val
+	}
 }
 
 func SetDefaults_PodExecOptions(obj *PodExecOptions) {