--- a/pkg/apis/apps/v1alpha1/zz_generated.defaults.go
+++ b/pkg/apis/apps/v1alpha1/zz_generated.defaults.go
@@ -80,6 +80,8 @@ func SetObjectDefaults_PetSet(in *PetSet) {
 				}
 			}
 		}
+		v1.SetDefaults_ResourceList(&a.Resources.Limits)
+		v1.SetDefaults_ResourceList(&a.Resources.Requests)
 		if a.LivenessProbe != nil {
 			v1.SetDefaults_Probe(a.LivenessProbe)
 			if a.LivenessProbe.Handler.HTTPGet != nil {
@@ -120,6 +122,8 @@ func SetObjectDefaults_PetSet(in *PetSet) {
 				}
 			}
 		}
+		v1.SetDefaults_ResourceList(&a.Resources.Limits)
+		v1.SetDefaults_ResourceList(&a.Resources.Requests)
 		if a.LivenessProbe != nil {
 			v1.SetDefaults_Probe(a.LivenessProbe)
 			if a.LivenessProbe.Handler.HTTPGet != nil {
@@ -148,6 +152,9 @@ func SetObjectDefaults_PetSet(in *PetSet) {
 	for i := range in.Spec.VolumeClaimTemplates {
 		a := &in.Spec.VolumeClaimTemplates[i]
 		v1.SetDefaults_PersistentVolumeClaim(a)
+		v1.SetDefaults_ResourceList(&a.Spec.Resources.Limits)
+		v1.SetDefaults_ResourceList(&a.Spec.Resources.Requests)
+		v1.SetDefaults_ResourceList(&a.Status.Capacity)
 	}
 }