--- a/pkg/apis/authentication/v1beta1/BUILD
+++ b/pkg/apis/authentication/v1beta1/BUILD
@@ -26,7 +26,6 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
         "//pkg/api/unversioned:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/apis/authentication:go_default_library",