--- a/pkg/apis/autoscaling/v1/BUILD
+++ b/pkg/apis/autoscaling/v1/BUILD
@@ -26,7 +26,6 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
         "//pkg/api/unversioned:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/apis/autoscaling:go_default_library",