--- a/pkg/apis/imagepolicy/v1alpha1/BUILD
+++ b/pkg/apis/imagepolicy/v1alpha1/BUILD
@@ -24,7 +24,6 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
         "//pkg/api/unversioned:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/apis/imagepolicy:go_default_library",