--- a/pkg/apis/rbac/v1alpha1/BUILD
+++ b/pkg/apis/rbac/v1alpha1/BUILD
@@ -26,7 +26,6 @@ go_library(
     ],
     tags = ["automanaged"],
     deps = [
-        "//pkg/api:go_default_library",
         "//pkg/api/unversioned:go_default_library",
         "//pkg/api/v1:go_default_library",
         "//pkg/apis/rbac:go_default_library",