… Godeps/Godeps.json
… cmd/libs/go2idl/conversion-gen/generators/conversion.go
… pkg/api/v1/defaults_test.go
… pkg/api/v1/zz_generated.conversion.go
… pkg/apis/authentication/v1beta1/zz_generated.conversion.go
… pkg/apis/authorization/v1beta1/zz_generated.conversion.go
… pkg/apis/autoscaling/v1/zz_generated.conversion.go
… pkg/apis/batch/v1/zz_generated.conversion.go
… pkg/apis/batch/v2alpha1/zz_generated.conversion.go
… pkg/apis/batch/v2alpha1/zz_generated.defaults.go
… pkg/apis/certificates/v1alpha1/zz_generated.conversion.go
… pkg/apis/componentconfig/v1alpha1/zz_generated.conversion.go
… pkg/apis/extensions/v1beta1/zz_generated.conversion.go
… pkg/apis/extensions/v1beta1/zz_generated.defaults.go
… pkg/apis/imagepolicy/v1alpha1/zz_generated.conversion.go
… pkg/apis/policy/v1alpha1/zz_generated.conversion.go
… pkg/apis/rbac/v1alpha1/zz_generated.conversion.go
… pkg/apis/storage/v1beta1/BUILD
… pkg/apis/storage/v1beta1/zz_generated.conversion.go
… plugin/pkg/scheduler/algorithm/priorities/interpod_affinity.go
… vendor/k8s.io/gengo/parser/parse.go