@smarterclayton 2016-10-14
Only last two commits are new, the others are part of the approved split conversion PR