@smarterclayton 2016-10-14
With some more tweaks, only a slight improvement here but should be larger in the real world.

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       3596          2340          -34.93%
BenchmarkPodDefaulting-8                       124           301           +142.74%
BenchmarkEncodeCodecFromInternalProtobuf-8     6034          4777          -20.83%
BenchmarkDecodeCodecToInternalProtobuf-8       7309          6418          -12.19%
BenchmarkEncodeCodecFromInternal-8             31265         31891         +2.00%
BenchmarkDecodeCodec-8                         65046         62757         -3.52%
BenchmarkDecodeIntoInternalCodec-8             61734         61801         +0.11%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       26             13             -50.00%
BenchmarkEncodeCodecFromInternalProtobuf-8     16             8              -50.00%
BenchmarkDecodeCodecToInternalProtobuf-8       72             64             -11.11%
BenchmarkEncodeCodecFromInternal-8             68             60             -11.76%
BenchmarkDecodeCodec-8                         268            260            -2.99%
BenchmarkDecodeIntoInternalCodec-8             268            260            -2.99%

benchmark                                      old bytes     new bytes     delta
BenchmarkPodConversion-8                       3324          2275          -31.56%
BenchmarkEncodeCodecFromInternalProtobuf-8     3992          3392          -15.03%
BenchmarkDecodeCodecToInternalProtobuf-8       5560          4960          -10.79%
BenchmarkEncodeCodecFromInternal-8             7180          6576          -8.41%
BenchmarkDecodeCodec-8                         13920         13320         -4.31%
BenchmarkDecodeIntoInternalCodec-8             13920         13320         -4.31%
```

I'm going to add a mechanism to treat certain manual conversions as no-op - any object that has private fields can more efficiently be copied via the unsafe pointer than via the innate mechanisms (things like unversioned.Time or resource.Quantity).  So not quite ready for merge but ready for review.