@smarterclayton 2016-10-14
With a few more changes, roughly the same in allocations.

```
BenchmarkPodConversion-8                     	  500000	      2384 ns/op	    2275 B/op	      13 allocs/op
BenchmarkNodeConversion-8                    	 1000000	      1742 ns/op	    1696 B/op	       9 allocs/op
BenchmarkReplicationControllerConversion-8   	  500000	      2465 ns/op	    2560 B/op	      15 allocs/op
BenchmarkPodCopy-8                           	 1000000	      1833 ns/op	    2000 B/op	      12 allocs/op
BenchmarkNodeCopy-8                          	 1000000	      1744 ns/op	    2192 B/op	       7 allocs/op
BenchmarkReplicationControllerCopy-8         	 1000000	      2335 ns/op	    2520 B/op	      15 allocs/op
BenchmarkPodDefaulting-8                     	 5000000	       299 ns/op	       0 B/op	       0 allocs/op
BenchmarkEncodeCodecProtobuf-8               	  500000	      3322 ns/op	    2288 B/op	       3 allocs/op
BenchmarkEncodeCodecFromInternalProtobuf-8   	  300000	      4792 ns/op	    3392 B/op	       8 allocs/op
BenchmarkEncodeProtobufGeneratedMarshal-8    	  500000	      2495 ns/op	    1088 B/op	       1 allocs/op
BenchmarkDecodeCodecToInternalProtobuf-8     	  200000	      6443 ns/op	    4960 B/op	      64 allocs/op
BenchmarkDecodeIntoProtobuf-8                	  300000	      3916 ns/op	    2784 B/op	      55 allocs/op
BenchmarkEncodeCodec-8                       	   50000	     28679 ns/op	    5466 B/op	      55 allocs/op
BenchmarkEncodeCodecFromInternal-8           	   50000	     30822 ns/op	    6575 B/op	      60 allocs/op
BenchmarkEncodeJSONMarshal-8                 	   50000	     28964 ns/op	    8080 B/op	      55 allocs/op
BenchmarkDecodeCodec-8                       	   20000	     59316 ns/op	   13320 B/op	     260 allocs/op
BenchmarkDecodeIntoExternalCodec-8           	   30000	     57126 ns/op	   12280 B/op	     255 allocs/op
BenchmarkDecodeIntoInternalCodec-8           	   20000	     61773 ns/op	   13320 B/op	     260 allocs/op
BenchmarkDecodeIntoJSON-8                    	   30000	     57158 ns/op	    7128 B/op	     110 allocs/op
BenchmarkDecodeIntoJSONCodecGen-8            	   50000	     33527 ns/op	    9336 B/op	     238 allocs/op
PASS
ok  	k8s.io/kubernetes/pkg/api	34.840s
○ benchcmp /tmp/old /tmp/new
benchmark                                      old ns/op     new ns/op     delta
BenchmarkPodConversion-8                       3596          2384          -33.70%
BenchmarkNodeConversion-8                      4493          1742          -61.23%
BenchmarkReplicationControllerConversion-8     3460          2465          -28.76%
BenchmarkPodDefaulting-8                       124           299           +141.13%
BenchmarkEncodeCodecFromInternalProtobuf-8     6034          4792          -20.58%
BenchmarkDecodeCodecToInternalProtobuf-8       7309          6443          -11.85%
BenchmarkEncodeCodecFromInternal-8             31265         30822         -1.42%
BenchmarkDecodeCodec-8                         65046         59316         -8.81%
BenchmarkDecodeIntoInternalCodec-8             61734         61773         +0.06%

benchmark                                      old allocs     new allocs     delta
BenchmarkPodConversion-8                       26             13             -50.00%
BenchmarkNodeConversion-8                      21             9              -57.14%
BenchmarkReplicationControllerConversion-8     23             15             -34.78%
BenchmarkEncodeCodecFromInternalProtobuf-8     16             8              -50.00%
BenchmarkDecodeCodecToInternalProtobuf-8       72             64             -11.11%
BenchmarkEncodeCodecFromInternal-8             68             60             -11.76%
BenchmarkDecodeCodec-8                         268            260            -2.99%
BenchmarkDecodeIntoInternalCodec-8             268            260            -2.99%

benchmark                                      old bytes     new bytes     delta
BenchmarkPodConversion-8                       3324          2275          -31.56%
BenchmarkNodeConversion-8                      4800          1696          -64.67%
BenchmarkReplicationControllerConversion-8     3456          2560          -25.93%
BenchmarkEncodeCodecFromInternalProtobuf-8     3992          3392          -15.03%
BenchmarkDecodeCodecToInternalProtobuf-8       5560          4960          -10.79%
BenchmarkEncodeCodecFromInternal-8             7180          6575          -8.43%
BenchmarkDecodeCodec-8                         13920         13320         -4.31%
BenchmarkDecodeIntoInternalCodec-8             13920         13320         -4.31%
```