@smarterclayton 2016-10-19
Depends on https://github.com/kubernetes/gengo/pull/12