@smarterclayton 2016-10-19
Spawned https://github.com/gogo/protobuf/pull/217 for the investigation of the other part of this - for our fat objects this is probably a net win (since we are string heavy).

Looks like some conversion bugs still exist, will debug more.